acq32/acq200: 64-bit kernel DMA fixes
The /proc/acq32 read panic on x86_64 traces to two distinct 64-bit
unsafety issues that were latent in the 2.4 source:

(1) acq32 DMA-buffer placement.  allocateDmaBuffer() called
__get_free_pages(GFP_KERNEL, ...).  On a 64-bit kernel with >4 GiB
of RAM that allocation can land above the 4 GiB line, which a 32-bit
DMA master (the DEC 21285) cannot reach.  The device DMAs to a
truncated address, hands back an MFA computed from the truncated
address, and the driver phys_to_virt()s its way to a random kernel
page — usually corrupting a slab object that cache_reap() then trips
over a few seconds later (a watchdog reboot is the visible symptom).

Two changes pin the buffer below 4 GiB:
* acq32_pci_probe() now sets pci_set_dma_mask(DMA_BIT_MASK(32)) and
  pci_set_consistent_dma_mask(DMA_BIT_MASK(32)).
* allocateDmaBuffer() ORs GFP_DMA32 into __get_free_pages so the
  allocation honours the mask even though it bypasses the dma_*
  coherent API.

(2) acq200_dev_i2o_mfa2va() truncated the kernel VA.  The original
expression was (void*)((unsigned)device->dmabuf.va + (mfa & MASK)) —
"unsigned" is 32-bit on x86_64, so a 64-bit kernel direct-map
pointer got chopped.  Use uintptr_t for the arithmetic.  This affects
the acq200-drv build only; acq32-drv reaches the same place via a
correct phys_to_virt() in acq32_dev_i2o_mfa2va().

Add #include <linux/dma-mapping.h> for DMA_BIT_MASK.

+++ project/BASE/DTACQ/SRC/dt100/Linux/Drivers/acq200.c
@@ -869,7 +869,11 @@ static void acq200_dev_i2o_return_mfa(struct Acq32Device* device, unsigned mfa)
 
 static void* acq200_dev_i2o_mfa2va(struct Acq32Device *device, u32 mfa)
 {
-	return (void*)((unsigned)device->dmabuf.va+(mfa&ACQ200_PCIWINMSK));
+	/* The 2.4 source cast dmabuf.va to (unsigned), which is 32-bit on
+	 * x86_64 — that truncated the kernel virtual address and the
+	 * resulting "VA" pointed at random low memory.  Use uintptr_t so
+	 * the pointer arithmetic survives 64-bit. */
+	return (void*)((uintptr_t)device->dmabuf.va + (mfa & ACQ200_PCIWINMSK));
 }
 static void acq200_incoming_message_isr(struct Acq32Device *device)
 {
